Add files via upload
## Summary of Findings

> Most Distributions were right skewed in the standard scale. Log scaling was performed to view the data properly. 
The transformed data all had normal distribution under log scaling and this makes sene that the most popular channels are a few portion of the total.
Some outliers made it very hard to examine the distributions so axis limits were imposed after examining the observations it seemed that the outliers are possible and no need for removing them. 
The ratio between the partnered and not partnered channels was very huge with only 22 not partnered observations making it hard to draw any useful conclusions.
Most correlations were weak and no obvious trends can be observed. The most interesting feature was the negative relationship between the streaming time and number of viewers.
I wasn't expecting the language to have a great effect on distribution of features but it yielded interesting results when compared for ratios of mature content.
The data set was not the best to work with, I think a bigger data set that includes genre of streamed content, year of joining Twitch would help draw far more powerful conclusions. 
The partnered variable had no significance as mostly all channels were partnered.